refactor: add request method in method name
