+++ functions/types.ts
@@ -1,5 +1,16 @@
 import { EventContext } from "@cloudflare/workers-types";
-import { Env } from "./shared";
+
+export interface Env {
+  USE_WIREX_SANDBOX: string;
+  WIREX_CLIENT_ID: string;
+  WIREX_CLIENT_SECRET: string;
+}
+
+export interface AccessToken {
+  token: string;
+  isSandbox: boolean;
+  expiresIn: number; // in seconds
+}
 
 export interface WirexAuthResponse {
   access_token: string;
