+++ functions/user-auth.ts
@@ -1,14 +1,11 @@
-import { validateRequestMethod } from "./register";
 import { getAccessToken } from "./shared";
+import { Context } from "./types";
 
-export async function onRequest(ctx): Promise<Response> {
+export async function onRequestPost(ctx: Context): Promise<Response> {
   try {
-    validateRequestMethod(ctx.request.method, "POST");
     const accessToken = await getAccessToken(ctx.env);
-    const result = await ctx.request.json();
-    // if (!result.success) {
-    //   throw new Error(`Invalid post parameters: ${JSON.stringify(result)}`);
-    // }
+    const result: { wallet: string } = await ctx.request.json();
+
     console.log("result", result);
     const { wallet } = result;
